requirements.txt
 openai>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
pathlib>=1.0.1
typing>=3.7.4
numpy>=1.24.0
logging>=0.4.9.6